Make it work with Jspm 0.16
By updating
* jspm-loader-css-modules to version 1.0.1-beta1
* jspm-loader-css to version 1.0.1-beta1

Note that due to a bug in SystemJS, one has to modify the file
jspm_packages/npm/postcss@5.0.10/lib/parse.js
so that a module isn't required in a string.
In other words, line 14 has to be changed to something like this
(just removing the require statement is enough)

throw new Error('Option safe was removed. ' + 'Use parser: "postcss-safe-parser"');

Fixes #4 #6 #7